requirements.txt
Django~=3.0.5
djangorestframework~=3.11.0
django-cors-headers~=3.2.1
django-filter~=2.2.0
djangorestframework-jwt==1.11.0
psycopg2-binary==2.8.6
djangorestframework-recursive
drf-schema-adapter
rest-pandas
drf-renderer-xlsx
drf-yasg